+++ lib/connected.dart
@@ -3,3 +3,5 @@ library connected;
 /// A Calculator.
 export 'src/app_connected.dart';
 export 'src/services/network_service.dart';
+export 'src/internet_checker/internet_connection.dart';
+export 'src/internet_checker/internet_check_option.dart';