+++ lib/src/configs/config.dart
@@ -1,0 +1,7 @@
+class Configs {
+  static const int defaultCheckInterval = 7;
+  static const int timeOut = 25;
+  static const int connectTimeout = timeOut;
+  static const int receiveTimeout = timeOut;
+  static const int sendTimeout = timeOut;
+}