+++ lib/src/services/network_service.dart
@@ -1,8 +1,10 @@
 import 'dart:async';
-
-import 'package:internet_connection_checker_plus/internet_connection_checker_plus.dart';
+import 'package:connected/src/configs/config.dart';
 
 import '../enums/network_status.dart';
+import '../internet_checker/internet_check_option.dart';
+import '../internet_checker/internet_connection.dart';
+import '../internet_checker/internet_status.dart';
 
 abstract class INetworkService {
   Future<NetworkStatus> check();
@@ -15,22 +17,28 @@ abstract class INetworkService {
 }
 
 class NetworkService extends INetworkService {
-  NetworkService() {
+  NetworkService({this.timeOut, this.checkInterval}) {
+    print("timeOut:   $timeOut");
+    print("checkInterval:   $checkInterval");
     _connection = InternetConnection.createInstance(
       useDefaultOptions: false,
-      checkInterval: const Duration(seconds: 5),
+      checkInterval: checkInterval ??
+          const Duration(seconds: Configs.defaultCheckInterval),
       customCheckOptions: [
-        InternetCheckOption(uri: Uri.parse('https://icanhazip.com')),
+        InternetCheckOption(
+            uri: Uri.parse('https://icanhazip.com'), timeout: timeOut)
       ],
     );
   }
 
   StreamSubscription<InternetStatus>? _subscription;
-  late InternetConnection _connection;
+  late InternetConnection? _connection;
+  Duration? timeOut;
+  Duration? checkInterval;
 
   @override
   Future<NetworkStatus> check() async {
-    bool result = await _connection.hasInternetAccess;
+    bool result = await _connection!.hasInternetAccess;
     return NetworkStatus.fromBool(result);
   }
 
@@ -41,11 +49,11 @@ class NetworkService extends INetworkService {
 
   @override
   void listenConnectivity(void Function(NetworkStatus status) onChange) {
-    _subscription = _connection.onStatusChange.listen((InternetStatus result) {
+    _subscription = _connection!.onStatusChange.listen((InternetStatus result) {
       onChange(NetworkStatus.fromInternetStatusResult(result));
     });
   }
 
   @override
-  Future<bool> get isConnected async => await _connection.hasInternetAccess;
+  Future<bool> get isConnected async => await _connection!.hasInternetAccess;
 }