+++ CHANGELOG.md
@@ -6,3 +6,8 @@ Flutter package for build real internet connection notifier on widget tree.
 ## 0.0.6+10
 
 Background black color problem when animation reversed solved.
+
+##  0.1.0
+
+- "connectionTimeOut" added for request response max time , must express by Duration()
+- "checkInterval" added for change of Connected widget accuracy